requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
astroid==2.15.6
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
charset-normalizer==3.1.0
colorama==0.4.6
DateTime==5.1
dill==0.3.7
discord==2.2.3
discord.py==2.2.3
emoji==2.4.0
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
logging==0.4.9.6
mccabe==0.7.0
multidict==6.0.4
Pillow==9.5.0
platformdirs==3.10.0
pylint==2.17.5
python-dotenv==1.0.0
pytz==2023.3
tomlkit==0.12.1
wrapt==1.15.0
yarl==1.9.2
zope.interface==6.0
